refactor(Participants): Dealer와 Player에게 상속해 주는 Participants 객체 구현
