+++ src/main/java/blackjack/domain/participants/Dealer.java
@@ -1,6 +1,6 @@
 package blackjack.domain.participants;
 
-public class Dealer extends Player {
+public class Dealer extends Participants {
 
     private static final int MAX_SUM_FOR_MORE_CARD = 16;
     private int winCount = 0;
