+++ src/main/java/blackjack/domain/participants/Participants.java
@@ -1,0 +1,41 @@
+package blackjack.domain.participants;
+
+import blackjack.domain.PlayerCards;
+import blackjack.domain.card.Card;
+import java.util.List;
+
+public class Participants {
+    private static final int BUST_LIMIT = 22;
+
+    private final String name;
+    private final PlayerCards playerCards;
+
+    public Participants(final String name) {
+        this.name = name;
+        this.playerCards = new PlayerCards();
+    }
+
+    public void receiveCard(final Card card) {
+        playerCards.receiveCard(card);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getCardCount() {
+        return playerCards.getCardCount();
+    }
+
+    public List<Card> getPlayerCards() {
+        return playerCards.toList();
+    }
+
+    public boolean isBust() {
+        return calculate() >= BUST_LIMIT;
+    }
+
+    public int calculate() {
+        return playerCards.calculate();
+    }
+}
