+++ src/main/java/blackjack/domain/participants/Player.java
@@ -1,44 +1,11 @@
 package blackjack.domain.participants;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.PlayerCards;
-import java.util.List;
+public class Player extends Participants {
 
-public class Player {
-
-    private static final int BUST_LIMIT = 22;
-
-    private final String name;
-    private final PlayerCards playerCards;
     private boolean win = true;
 
-    public Player(final String name) {
-        this.name = name;
-        this.playerCards = new PlayerCards();
-    }
-
-    public void receiveCard(final Card card) {
-        playerCards.receiveCard(card);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getCardCount() {
-        return playerCards.getCardCount();
-    }
-
-    public List<Card> getPlayerCards() {
-        return playerCards.toList();
-    }
-
-    public boolean isBust() {
-        return calculate() >= BUST_LIMIT;
-    }
-
-    public int calculate() {
-        return playerCards.calculate();
+    public Player(String name) {
+        super(name);
     }
 
     public void lose() {
