+++ src/test/java/blackjack/DealerTest.java
@@ -7,6 +7,7 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardType;
 import blackjack.domain.participants.Dealer;
+import blackjack.domain.participants.Participants;
 import blackjack.domain.participants.Player;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.AssertionsForClassTypes;
@@ -35,9 +36,9 @@ public class DealerTest {
     @Test
     @DisplayName("딜러가 참가자에게 상속받았는지 확인")
     void extend() {
-        final Player player = new Dealer("딜러");
-        player.receiveCard(new Card(CardNumber.ACE, CardType.HEART));
-        assertThat(player.getCardCount()).isEqualTo(1);
+        final Participants participants = new Dealer("딜러");
+        participants.receiveCard(new Card(CardNumber.ACE, CardType.HEART));
+        assertThat(participants.getCardCount()).isEqualTo(1);
     }
 
     @Test
